test: AdminPageControllerTest와 UserPageControllerTest를 WebMvcConfigTest로 통합
