+++ src/test/java/roomescape/global/config/WebMvcConfigTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.global.config;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -8,7 +8,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class AdminPageControllerTest {
+class WebMvcConfigTest {
 
     @Test
     @DisplayName("ADMIN 예약 관리 메인 페이지를 렌더링한다")
@@ -42,6 +42,33 @@ class AdminPageControllerTest {
     void displayAdminThemePage() {
         RestAssured.given().log().all()
                 .when().get("admin/theme")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("User 메인 페이지를 렌더링한다")
+    void displayMainPage() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("User 예약 목록 페이지를 렌더링한다")
+    void displayReservationPage() {
+        RestAssured.given().log().all()
+                .when().get("reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("로그인 페이지를 렌더링한다")
+    void displayLoginPage() {
+        RestAssured.given().log().all()
+                .when().get("login")
                 .then().log().all()
                 .statusCode(200);
     }
